+++ packages/components/pages/f-checkout/test/visual/f-checkout-invalid.visual.mobile.spec.js
@@ -34,6 +34,7 @@ describe('f-checkout - Invalid - Mobile Visual Tests', () => {
     });
 
     it('should display the error page component if the url is invalid', () => {
+        // Arrange
         checkout = new Checkout();
 
         const args = argumentStringBuilder({ serviceType: 'invalid-url' });
@@ -46,6 +47,7 @@ describe('f-checkout - Invalid - Mobile Visual Tests', () => {
     });
 
     it('should display an error dialog if the basket has invalid products', () => {
+        // Arrange
         checkout = new Checkout();
 
         const args = argumentStringBuilder({ getBasketError: 'invalid-products' });
@@ -59,6 +61,7 @@ describe('f-checkout - Invalid - Mobile Visual Tests', () => {
     });
 
     it('should display an error dialog if the basket has offline products', () => {
+        // Arrange
         checkout = new Checkout();
 
         const args = argumentStringBuilder({ getBasketError: 'offline-products' });
@@ -72,6 +75,7 @@ describe('f-checkout - Invalid - Mobile Visual Tests', () => {
     });
 
     it('should display an error dialog if age verification is required', () => {
+        // Arrange
         checkout = new Checkout();
 
         const args = argumentStringBuilder({ getBasketError: 'age-restriction' });