+++ packages/components/organisms/f-cookie-banner/test-utils/component-objects/f-cookie-banner.component.js
@@ -3,15 +3,15 @@ const Page = require('@justeat/f-wdio-utils/src/page.object');
 module.exports = class CookieBanner extends Page {
     get component () { return $('[data-test-id="cookieBanner-component"]'); }
 
-    open (url) {
-        super.open(url);
+    async open (url) {
+        await super.open(url);
     }
 
-    waitForComponent () {
-        this.component.waitForExist();
+    async waitForComponent () {
+        await this.component.waitForExist();
     }
 
-    isCookieBannerComponentDisplayed () {
+    async isCookieBannerComponentDisplayed () {
         return this.component.isDisplayed();
     }
 };